+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- wrong diff behavior when updating entities (unsavedDiff())
 
 
 ## [0.5.0] - 2025-03-18